WIP to refactor project plotting functionality
