+++ .gitignore
@@ -97,5 +97,5 @@ ENV/
 
 # Docathon ignored pages
 blog/content/pages/projects/*
-src/watchtower/.*
 src/data/*
+src/.users.csv
