+++ src/watchtower/.downloaded_projects
@@ -1,0 +1,21 @@
+pmagpy,pmagpy
+gallantlab,pycortex
+gallantlab,cottoncandy
+data-8,connector-instructors
+matplotlib,matplotlib
+jupyter,notebook
+jmschrei,pomegranate
+pandas-dev,pandas
+yt_analysis,yt
+scikit-image,scikit-image
+clawpack,clawpack
+mne-tools,mne-python
+scikit-learn-contrib,hdbscan
+scikit-learn,scikit-learn
+Sylius,Sylius
+omeka,Omeka
+ROBelgium,MSNoise
+kbenoit,readtext
+kbenoit,quanteda
+ipython,ipython
+bokeh,bokeh
