+++ src/watchtower/calculate_project_commits.py
@@ -1,15 +1,12 @@
 import numpy as np
 import os
-import matplotlib
-matplotlib.use('agg')
-import matplotlib.dates as mpd
 from watchtower import commits_
-import matplotlib.pyplot as plt
 from tqdm import tqdm
-import calendar
 import pandas as pd
 import traceback
 
+# THIS SHOULD BE BROKEN UP INTO 'CALCULATE_PROJECT_COMMITS' and "PLOT_COMMITS"
+# CALCULATE WOULD BE RUN ON THE USER SIDE, and PLOT WOULD BE RUN ON THE SERVER
 
 def count_doc_commits(user, project, search_queries=None,
                       groupby='month', start='2017-01-01', stop=None):
@@ -37,8 +34,6 @@ def count_doc_commits(user, project, search_queries=None,
     else:
         stop = pd.to_datetime(stop)
 
-    if project == 'IPython':
-        import IPython; IPython.embed()
     commits = commits_.load_commits(user, project)
     if commits is None:
         return None, None
@@ -59,6 +54,12 @@ def count_doc_commits(user, project, search_queries=None,
     # Remove commits from the past we don't want
     mask_since = (dates > start) * (dates < stop)
     commits = commits[mask_since]
+    if len(commits) == 0:
+        # In case no commits for this project
+        all_dates = all_dates.drop(0, axis=1).astype(int)
+        all_dates['all'] = 0
+        all_dates['doc'] = 0
+        return all_dates
 
     # Find commits that match our queries
     mask_doc = np.zeros(commits.shape[0])
@@ -75,48 +76,15 @@ def count_doc_commits(user, project, search_queries=None,
     commits_all = commits['is_doc'].resample('D').count()
 
     for date, val in commits_all.items():
-        all_dates.loc[date, 'All'] = val
+        all_dates.loc[date, 'all'] = val
     for date, val in commits_doc.items():
-        all_dates.loc[date, 'Doc'] = val
+        all_dates.loc[date, 'doc'] = val
 
     # Clean up
     all_dates = all_dates.drop(0, axis=1)
     all_dates = all_dates.replace(np.nan, 0)
     all_dates = all_dates.astype(int)
     return all_dates
-
-
-def plot_commits(all_dates):
-
-    # --- Plotting ---
-    fig, ax = plt.subplots(figsize=(8, 4))
-    for label in all_dates.columns:
-        ax.bar(all_dates.index.to_pydatetime(), all_dates[label].values,
-               label=label)
-    # Plot today
-    today = pd.datetime.today()
-    ax.axvline(today, ls='--', alpha=.5, lw=2, color='k')
-    ax.grid("off")
-    ax.spines['right'].set_color('none')
-    ax.spines['left'].set_color('none')
-    ax.spines['top'].set_color('none')
-    ax.xaxis.set_ticks_position('bottom')
-    ax.yaxis.set_ticks_position('left')
-    ax.xaxis.set_major_formatter(mpd.DateFormatter('%b\n%d'))
-
-    # Y-axis formatting
-    ax.set_ylabel("# commits")
-    ax.set_ylim([0, np.max([5, int(ax.get_ylim()[-1])])])
-    yticks = ax.get_yticks()
-    for l in yticks:
-        ax.axhline(l, linewidth=0.75, zorder=-10, color="0.5")
-    ax.set_yticks(yticks)
-
-    ax.legend(loc=1)
-    ax.set_title(project, fontweight="bold", fontsize=22)
-    plt.tight_layout()
-    plt.autoscale(tight=True)
-    return fig, ax
 
 
 # --- Run the script ---
@@ -127,20 +95,14 @@ except OSError:
     pass
 
 groupby = 'weekday'
-start = '2017-02-02'
+start = '2017-03-01'
 stop = '2017-03-10'
 exceptions = []
 all_dates = []
 for user, project in tqdm(informations):
     try:
-        this_all_dates = count_doc_commits(user, project,
-                                      groupby=groupby, start=start, stop=stop)
-        fig, ax = plot_commits(this_all_dates)
-        if fig is None:
-            exceptions.append(project)
-            continue
-        filename = os.path.join("build/images", project.lower() + ".png")
-        fig.savefig(filename, bbox_inches='tight')
+        this_all_dates = count_doc_commits(
+            user, project, groupby=groupby, start=start, stop=stop)
 
         # Collect data so we can save it
         this_all_dates['project'] = project
@@ -148,6 +110,8 @@ for user, project in tqdm(informations):
     except Exception as e:
         exceptions.append(project)
         traceback.print_exception(None, e, e.__traceback__)
+
+print('Exceptions: {}'.format(exceptions))
 all_dates = pd.concat(all_dates, axis=0)
+all_dates.index.name = 'date'
 all_dates.to_csv('.totals.csv')
-print('Finished building images.\nExceptions: {}'.format(exceptions))
